Filter entities folder renamed and index added
