+++ generator/entities/filter/index.js
@@ -1,0 +1,3 @@
+export { default as Block } from './block.js';
+export { default as Filter } from './filter.js';
+export { default as Section } from './section.js';
