+++ generator/entities/filter/section.js
