+++ generator/sections/900-salvageables.js
@@ -1,4 +1,4 @@
-import Section from '../entities/filter-components/section.js';
+import { Section } from '../entities/filter/index.js';
 import MapIcon from '../entities/map-icon.js';
 import Card from '../entities/card.js';
 import { global, modes } from '../configuration/index.js';
