+++ generator/sections/902-useless.js
@@ -1,4 +1,4 @@
-import Section from '../entities/filter-components/section.js';
+import { Section } from '../entities/filter/index.js';
 import { global, modes } from '../configuration/index.js';
 
 export default ({ modeId }) => {
