+++ generator/sections/999-debug.js
@@ -1,4 +1,4 @@
-import Section from '../entities/filter-components/section.js';
+import { Section } from '../entities/filter/index.js';
 import Sound from '../entities/sound.js';
 import MapIcon from '../entities/map-icon.js';
 import Card from '../entities/card.js';
